// src/interfaces/projects.interface.tsx
export interface IProjects {
  id: number;
  name: string;
  description: string;
  image: string;
  url: string;
  github: string;
  technologies: string[];
  data: string;
}

// card

export interface ICard {
  name: string;
  description: string;
  image: string;
  url: string;
  github: string;
  technologies: string[];
  data: string;
}